+++ apibase/graphql/mixins.py
@@ -1,6 +1,11 @@
+import json
+
 import graphene.relay
+from django.db.models import QuerySet
+from graphene.types import generic
 
 from .. import serializers
+from .encoders import JSONEncode
 
 
 class NodeMixin(object):
@@ -33,3 +38,25 @@ class NodeMixin(object):
             return queryset.get(pk=id)
         except cls._meta.model.DoesNotExist:
             return None
+
+
+class SummaryMixin(object):
+    total_count = graphene.Int()
+    records = graphene.Int()
+    summary = generic.GenericScalar()
+
+    def resolve_total_count(self, info, **kwargs):
+        return self.length
+
+    def resolve_summary(self, info, **kwargs):
+        if isinstance(self.iterable, QuerySet) and hasattr(self.iterable, "summary"):
+            data = json.dumps(self.iterable.summary(), cls=JSONEncode)
+            return json.loads(data)
+        return None
+
+    def resolve_records(self, info, **kwargs):
+        if isinstance(self.iterable, QuerySet):
+            # TODO: each models may have it own countable criteria
+            return self.iterable.order_by("id").distinct().count()
+
+        return self.length
